Renamed similar_columns as match_columns
