+++ feature_engine/preprocessing/__init__.py
@@ -2,7 +2,7 @@
 The module preprocessing includes classes to check that train and tests are similar.
 """
 
-from .similar_columns import SimilarColumns
+from .match_columns import SimilarColumns
 
 __all__ = [
     "SimilarColumns",
